+++ src/FluxFlow.Components.FileSystem/Timing/IFileSystemClock.cs
@@ -1,0 +1,6 @@
+namespace FluxFlow.Components.FileSystem.Timing;
+
+public interface IFileSystemClock
+{
+    DateTimeOffset UtcNow { get; }
+}